+++ apps/libs/exception.py
@@ -1,20 +1,21 @@
 import traceback
 from typing import Any
 
-from fastapi import FastAPI
-from fastapi import Request, status
+from fastapi import FastAPI, Request, status
 from fastapi.exceptions import HTTPException, RequestValidationError
 from fastapi.responses import JSONResponse
 
-from extensions import StatusCode, BaseHTTPException, logger
+from conf.const import StatusCode
+from extensions.log import logger
+from extensions.exceptions import BaseHTTPException
 
 
-def log_message(request: Request, message: Any):
+def log_message(method: str, url: str, message: Any):
     """log message when catch exception"""
 
-    logger.error('start error'.center(60, '*'))
-    logger.error(f'{request.method} {request.url}')
-    logger.error(f'error is {message}')
+    logger.error('start error, this is'.center(60, '*'))
+    logger.error(f'{method} {url}')
+    logger.error(message)
     logger.error('end error'.center(60, '*'))
 
 
@@ -23,25 +24,25 @@ def register_exception(app: FastAPI):
     async def catch_c_http_exception(request: Request, exc: BaseHTTPException):
         """catch custom exception"""
 
-        log_message(request, exc.message)
-        content = {'status_code': exc.CODE, 'message': exc.message, 'data': None}
+        log_message(request.method, request.url, exc.message)
+        content = exc.response()
         return JSONResponse(content=content, status_code=exc.status_code, headers=exc.headers)
 
     @app.exception_handler(HTTPException)
     async def http_exception_handler(request: Request, exc: HTTPException):
         """catch FastAPI HTTPException"""
 
-        log_message(request, exc.detail)
-        content = {'status_code': StatusCode.bad_request, 'message': exc.detail, 'data': None}
+        log_message(request.method, request.url, exc.detail)
+        content = {'code': StatusCode.bad_request, 'message': exc.detail, 'data': None}
         return JSONResponse(content=content, status_code=exc.status_code, headers=exc.headers)
 
     @app.exception_handler(AssertionError)
     async def assert_exception_handle(request: Request, exc: AssertionError):
         """catch Python AssertError"""
 
-        exc_str = ''.join(exc.args)
-        log_message(request, exc_str)
-        content = {'status_code': StatusCode.validator_error, 'message': exc_str, 'data': None}
+        exc_str = ' '.join(exc.args)
+        log_message(request.method, request.url, exc_str)
+        content = {'code': StatusCode.validator_error, 'message': exc_str, 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     @app.exception_handler(RequestValidationError)
@@ -49,15 +50,15 @@ def register_exception(app: FastAPI):
         """catch FastAPI RequestValidationError"""
 
         exc_str = f'{exc}'.replace('\n', ' ').replace('   ', ' ')
-        log_message(request, exc_str)
+        log_message(request.method, request.url, exc)
         # content = exc.errors()
-        content = {'status_code': StatusCode.validator_error, 'message': exc_str, 'data': None}
+        content = {'code': StatusCode.validator_error, 'message': exc_str, 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     @app.exception_handler(Exception)
     async def exception_handle(request: Request, exc: Exception):
         """catch other exception"""
 
-        log_message(request, traceback.format_exc())
-        content = {'status_code': StatusCode.server_error, 'message': str(exc), 'data': None}
+        log_message(request.method, request.url, traceback.format_exc())
+        content = {'code': StatusCode.server_error, 'message': str(exc), 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)