+++ extensions/response.py
@@ -2,6 +2,10 @@ from typing import Any
 
 from extensions import logger
 
+from conf.const import StatusCode
+
+
+
 
 def resp_success(message: str = '', print_msg: str = '', data: Any = None):
     if print_msg:
@@ -13,4 +17,4 @@ def resp_success(message: str = '', print_msg: str = '', data: Any = None):
     if print_msg:
         logger.info(print_msg)
 
-    return {'status_code': 10000, 'message': message, 'data': data}
+    return {'message': message, 'data': data}